django/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile
#
amqp==5.0.6
    # via kombu
asgiref==3.4.1
    # via
    #   django
    #   uvicorn
astroid==2.6.2
    # via pylint
backcall==0.2.0
    # via ipython
billiard==3.6.4.0
    # via celery
cairocffi==1.2.0
    # via
    #   cairosvg
    #   weasyprint
cairosvg==2.5.2
    # via weasyprint
celery==5.1.2
    # via -r requirements.in
certifi==2021.5.30
    # via sentry-sdk
cffi==1.14.5
    # via
    #   cairocffi
    #   weasyprint
click==7.1.2
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   uvicorn
click-didyoumean==0.0.3
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.2.0
    # via celery
cssselect2==0.4.1
    # via
    #   cairosvg
    #   weasyprint
decorator==5.0.9
    # via
    #   ipdb
    #   ipython
defusedxml==0.7.1
    # via cairosvg
django==3.2.5
    # via
    #   -r requirements.in
    #   django-braces
    #   django-cors-headers
    #   django-extensions
    #   django-filter
    #   django-model-controller
    #   django-sequences
    #   djangorestframework
    #   djangorestframework-simplejwt
django-braces==1.14.0
    # via django-model-controller
django-cors-headers==3.7.0
    # via -r requirements.in
django-crispy-forms==1.12.0
    # via django-model-controller
django-environ-2==2.1.0
    # via -r requirements.in
django-excel-tools==1.1.0
    # via -r requirements.in
django-extensions==3.1.3
    # via -r requirements.in
django-filter==2.4.0
    # via -r requirements.in
django-model-controller==0.4.2
    # via -r requirements.in
django-sequences==2.6
    # via -r requirements.in
djangorestframework==3.12.4
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
djangorestframework-simplejwt==4.7.2
    # via -r requirements.in
et-xmlfile==1.1.0
    # via openpyxl
factory-boy==3.2.0
    # via -r requirements.in
faker==8.9.1
    # via factory-boy
gunicorn==20.1.0
    # via -r requirements.in
h11==0.12.0
    # via uvicorn
html5lib==1.1
    # via weasyprint
httptools==0.2.0
    # via -r requirements.in
ipdb==0.13.9
    # via -r requirements.in
ipython==7.25.0
    # via ipdb
ipython-genutils==0.2.0
    # via traitlets
isort==5.9.1
    # via pylint
jedi==0.18.0
    # via ipython
kombu==5.1.0
    # via celery
lazy-object-proxy==1.6.0
    # via astroid
matplotlib-inline==0.1.2
    # via ipython
mccabe==0.6.1
    # via pylint
openpyxl==3.0.7
    # via -r requirements.in
parso==0.8.2
    # via jedi
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pillow==8.3.1
    # via
    #   -r requirements.in
    #   cairosvg
    #   weasyprint
prompt-toolkit==3.0.19
    # via
    #   click-repl
    #   ipython
psycopg2-binary==2.9.1
    # via -r requirements.in
ptyprocess==0.7.0
    # via pexpect
pycparser==2.20
    # via cffi
pygments==2.9.0
    # via ipython
pyjwt==2.1.0
    # via djangorestframework-simplejwt
pylint==2.9.3
    # via -r requirements.in
pymupdf==1.18.14
    # via -r requirements.in
pyphen==0.11.0
    # via weasyprint
python-dateutil==2.8.1
    # via faker
pytz==2021.1
    # via
    #   -r requirements.in
    #   celery
    #   django
redis==3.5.3
    # via -r requirements.in
sentry-sdk==1.2.0
    # via -r requirements.in
six==1.16.0
    # via
    #   click-repl
    #   django-braces
    #   html5lib
    #   python-dateutil
sqlparse==0.4.1
    # via django
tblib==1.7.0
    # via -r requirements.in
text-unidecode==1.3
    # via faker
tinycss2==1.1.0
    # via
    #   cairosvg
    #   cssselect2
    #   weasyprint
toml==0.10.2
    # via
    #   ipdb
    #   pylint
traitlets==5.0.5
    # via
    #   ipython
    #   matplotlib-inline
unittest-xml-reporting==3.0.4
    # via -r requirements.in
urllib3==1.26.6
    # via sentry-sdk
uvicorn==0.14.0
    # via -r requirements.in
uvloop==0.15.2
    # via -r requirements.in
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.5
    # via prompt-toolkit
weasyprint==52.5
    # via -r requirements.in
webencodings==0.5.1
    # via
    #   cssselect2
    #   html5lib
    #   tinycss2
wrapt==1.12.1
    # via astroid

# The following packages are considered to be unsafe in a requirements file:
# setuptools
